# resources/openapi.yaml
openapi: 3.0.0
info:
  title: SciQuel Backend API
  version: 1.0.0
  description: Backend API for SciQuel Web Application
paths:
...
  /api/stories/search:
    get:
      tags:
        - Stories
      summary: Return a list of stories based on search query
      description: Return a list of stories that matches to search query
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
          required: false
          description: Page number
        - in: query
          name: page_size
          schema:
            type: integer
            minimum: 1
            default: 10
          required: false
          description: Number of stories per page
        - in: query
          name: keyword
          schema:
            type: string
          required: false
          description: Filter by keyword on story title, summary
        - in: query
          name: staff_pick
          schema:
            type: boolean
            default: false
          required: false
          description: Return only staff pick stories
        - in: query
          name: topic
          schema:
            type: string
            example: "ASTRONOMY"
          required: false
          description: Filter stories by topic (non-case sensitive)
        - in: query
          name: type
          schema:
            type: string
            example: "ARTICLE"
          required: false
          description: Filter stories by storyType (non-case sensitive)
        - in: query
          name: date_from
          schema:
            type: string
            example: "2023-06-13"
          required: false
          description: Filter stories after or on this date
        - in: query
          name: date_to
          schema:
            type: string
            example: "2023-06-13"
          required: false
          description: Filter stories before or on this date
        - in: query
          name: sort_by
          schema:
            type: string
            example: "newest | oldest"
          required: false
          description: Sort stories with preset rules
      responses:
        "200":
          description: Successfully return a list of stories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Story"
        "400":
          description: User requested invalid page or page size or topic
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Bad Request
...
components:
  schemas:
    Story:
      type: object
      properties:
        id:
          type: string
        storyType:
          $ref: "#/components/schemas/StoryType"
        title:
          type: string
        titleColor:
          type: string
        slug:
          type: string
        summary:
          type: string
        summaryColor:
          type: string
        tags:
          type: array
          items:
            $ref: "#/components/schemas/StoryTopic"
        storyContributions:
          type: array
          items:
            $ref: "#/components/schemas/StoryContribution"
        storyContent:
          type: array
          items:
            $ref: "#/components/schemas/StoryContent"
        published:
          type: boolean
        staffPick:
          type: boolean
        thumbnailUrl:
          type: string
        createdAt:
          type: string
          format: date-time
        publishedAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    StoryType:
      type: string
      enum:
        - ARTICLE
    StoryTopic:
      type: string
      enum:
        - ASTRONOMY
        - BIOLOGY
        - CHEMICAL_ENGINEERING
        - CHEMISTRY
        - COMPUTER_SCIENCE
        - ELECTRICAL_ENGINEERING
        - ENVIRONMENTAL_SCIENCE
        - GEOLOGY
        - MATHEMATICS
        - MECHANICAL_ENGINEERING
        - MEDICINE
        - PHYSICS
        - PSYCHOLOGY
        - SOCIOLOGY
        - TECHNOLOGY
    StoryContribution:
      type: object
      properties:
        id:
          type: string
        user:
          $ref: "#/components/schemas/User"
        userId:
          type: string
        story:
          $ref: "#/components/schemas/Story"
        storyId:
          type: string
        contributionType:
          $ref: "#/components/schemas/ContributionType"
    StoryContent:
      type: object
      properties:
        id:
          type: string
        story:
          $ref: "#/components/schemas/Story"
        storyId:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
    User:
      type: object
      properties:
        id:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        bio:
          type: string
        passwordHash:
          type: string
        joinedAt:
          type: string
          format: date-time
        storyContributions:
          type: array
          items:
            $ref: "#/components/schemas/StoryContribution"
    ContributionType:
      type: string
      enum:
        - AUTHOR
        - ILLUSTRATOR
        - ANIMATOR
...